SDK-537: Move attribute tests into separate files in attribute subpackage
