+++ test/fixtures/test_third_party_issuance_details.txt
